PayFragment.
Renamed layoutView to layoutPay.
Create Buttons for navigation and ImageButtons for coins.
Register nClickListener for all theses buttons.
On „abort“ pressed return to OverviewFragment.
On „Quittung“ pressed show TOAST with TODO, need to create bill.
For each press on a coin images a TOAST with the correct value is
shown. Need to reduce the money left to pay every time!
Later a data structure to handle the amount of money to pay is needed!